added VE EOS on the GPU
