+++ sph/include/sph/hydro_ve/CMakeLists.txt
@@ -1,6 +1,7 @@
 
 set(VE_KERNELS
         xmass_gpu.cu
+        eos_gpu.cu
         ve_def_gradh_gpu.cu
         iad_divv_curlv_gpu.cu
         av_switches_gpu.cu
