+++ sph/include/sph/sph.hpp
@@ -13,6 +13,7 @@
 #include "sph/hydro_std/momentum_energy.hpp"
 
 #include "sph/hydro_ve/av_switches.hpp"
+#include "sph/hydro_ve/eos.hpp"
 #include "sph/hydro_ve/ve_def_gradh.hpp"
 #include "sph/hydro_ve/iad_divv_curlv.hpp"
 #include "sph/hydro_ve/momentum_energy.hpp"
